refactor: enhance game processing and analysis configuration
- Introduced a new `analysis-config.ts` module to normalize analysis configurations and streamline game processing.
- Updated `GameProcessor` to utilize the new normalization logic, improving clarity and maintainability.
- Added `tracker-merge.ts` for handling worker results and merging tracker data, enhancing performance tracking.
- Created `games-to-pgn.ts` for re-encoding game objects as PGN text with optional header inclusion.
- Updated `AGENTS.md` to reflect changes in module structure and functionality.
